fixing hooks
