+++ Jenkinsfile_CNP
@@ -10,7 +10,7 @@ properties([
         string(name: 'CCD_GATEWAY_URL', defaultValue: 'https://gateway-ccd.aat.platform.hmcts.net', description: 'The CCD Gateway URL these tests will need to run'),
         string(name: 'TEST_ENVIRONMENT', defaultValue: 'aat', description: 'The environment to pull azure keys from'),
         string(name: 'NUM_TESTS_IN_PARALLEL', defaultValue: '4', description: 'The number of tests to run in parallel'),
-        string(name: 'RETRIES', defaultValue: '0', description: 'The number of retries to set for tests'),
+        string(name: 'RETRIES', defaultValue: '4', description: 'The number of retries to set for tests'),
         string(name: 'SPECIFIC_TAGS', defaultValue: '', description: 'Tag for a specific run of tests'),
     ])
 ])
